Split core and fix dependencies
According to the eslint documentation for custom configs, parsers and
configs you depend on belong in the `dependencies` and plugins belong
in the `peerDependencies`. This means that for npm users before v7 they
must manually install our peerDependencies. I'll this to the
documentation (cribbed from the docs for the airbnb config)

We now provide three configs:
- `eda/core` the smallest set
- `eda` core + jest
- `eda/react` core + jest + react

This covers all our existing challenges, in the future we might need one
that provides react but not jest ...