Replace eslint-plugin-prettier/recommended
We've moved from using eslint-plugin-prettier (+ recommended ruleset) to
just using `eslint-config-prettier`. This means that eslint is
configured to avoid conflicts with prettier, but also it doesn't enforce
prettier standards as an eslint rule.

Enforcing prettier standard as an eslint rule isn't necessary:

1. everyone should be using `editor.formatOnSave` or formatting from the
commandline
2. we can use `npx prettier --check` or similar to enforce

My main motivation behind using a formatter is that we can stop thinking
about formatting, not that we get really good at it. Some of the
students are finding it stressful that their lines are underlined in red
as they type.

With this change, they don't get the squiggly lines as they type, but
as soon as they save, their formatting is corrected